+++ changelog.md
@@ -1,1 +1,1 @@
-unbreakable tag
+- placing a target block underneath the dummy will output power when the dummy is hit